+++ Src/SALTED_utilities.cpp
@@ -233,7 +233,7 @@ cvec4 Rascaline_Descriptors::get_expansion_coeffs(std::vector<uint8_t> descripto
 
     metatensor::TensorMap descriptor = metatensor::TensorMap::load_buffer(descriptor_buffer);
     // cvec4 omega(this->nang + 1, std::vector<cvec2>(this->n_atoms, cvec2(2 * this->nang + 1, cvec(this->nspe * this->nrad, {0.0, 0.0}))));
-    cvec4 omega(this->n_atoms, std::vector<cvec2>(this->nspe * this->nrad, cvec2(this->nang + 1, cvec(2 * this->nang + 1, {0.0, 0.0}))));
+    cvec4 omega(this->n_atoms, std::vector<cvec2>((size_t)this->nspe * this->nrad, cvec2((size_t)this->nang + 1, cvec((size_t)2 * this->nang + 1, {0.0, 0.0}))));
     for (int l = 0; l < nang + 1; ++l)
     {
         cvec2 c2r = SALTED_Utils::complex_to_real_transformation({(2 * l) + 1})[0];
